PartnerChains.Scripts.Versioning

Description

Version Oracle Token minting policy

This is part of Partner Chains' versioning system. The versioning system stores ScriptIds with their corresponding reference scripts. This creates indirection that allows replacing scripts in the versioning system with different versions of them.

Versioned scripts can be looked up using getVersionedCurrencySymbol and getVersionedScriptHash.

Version Oracle Tokens identify authorized versioning UTXOs.

Redeemers:

InitializeVersionOracle initializes the first VersionOracle UTXO.

TX has to use the Genesis UTXO as an input.

TX must mint a single Version Oracle Token.

There must exist exactly one output at the validator address that carries:

VersionOracleDatum as datum matching script argument,

a script reference matching script argument,

and a single Version Oracle Token.

MintVersionOracle creates a new VersionOracle UTXO.

It is a governance action (see approvedByGovernance).

TX must mint a single Version Oracle Token.

There must exist exactly one output at the validator address that carries:

VersionOracleDatum as datum matching script argument,

a script reference matching script argument,

and a single Version Oracle Token.

BurnVersionOracle spends a VersionOracle UTXO.

It is a governance action (see approvedByGovernance).

There must exist exactly one input at the validator address that carries:

VersionOracleDatum as datum matching script argument,

and a single Version Oracle Token.

There must NOT exist any outputs carrying any Version Oracle Tokens.

Error codes:

ERROR-VERSION-POLICY-01: Genesis UTXO was not used as input.

ERROR-VERSION-POLICY-02: There is not a single output at the validator address with correct VersionOracleDatum, reference script, and a single version oracle token.

ERROR-VERSION-POLICY-03: One versioning token was not minted.

ERROR-VERSION-POLICY-04: There is not a single output at the validator address with correct VersionOracleDatum, reference script, and a single version oracle token.

ERROR-VERSION-POLICY-05: Transaction should be signed by the governance.

ERROR-VERSION-POLICY-06: One versioning token was not minted.

ERROR-VERSION-POLICY-07: There is not a single input at the validator address with expected old VersionOracleDatum, and a single version oracle token.

ERROR-VERSION-POLICY-08: There was an output carrying version oracle tokens.

ERROR-VERSION-POLICY-09: Transaction should be signed by the governance.

ERROR-VERSION-POLICY-10: Script purpose is not Minting.

mkVersionOraclePolicy :: TxOutRef -> Address -> VersionOraclePolicyRedeemer -> ScriptContext -> Bool

mkVersionOraclePolicyUntyped

Arguments

:: BuiltinData
Genesis Utxo

-> BuiltinData
Validator address

-> BuiltinData
Redeemer

-> BuiltinData
ScriptContext

-> BuiltinUnit

compiledVersionOraclePolicy :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> BuiltinData -> BuiltinUnit)

serialisableVersionOraclePolicy :: SerialisedScript

Version Oracle Token validator

Stores VersionOraclePolicy UTXOs, acting both as an oracle of available scripts as well as a script caching system. UTXOs on the script are managed by the governance authority, with VersionOraclePolicy ensuring that tokens are minted and burned correctly.

Spending from the VersionOracle validator is a governance action (see approvedByGovernance).

VersionOracle in redeemer must match the one in the datum.

Transaction must not transfer any Version Oracle tokens to another address.

Error codes:

ERROR-VERSION-VALIDATOR-01: Governance approval is not present

ERROR-VERSION-VALIDATOR-02: Version oracle in the datum does not match the redeemer

ERROR-VERSION-VALIDATOR-03: Transaction outputs version tokens to non-versioning address

ERROR-VERSION-VALIDATOR-04: Invalid script purpose is provided

ERROR-VERSION-VALIDATOR-05: Transaction does not have own input

mkVersionOracleValidator :: BuiltinData -> VersionOracleDatum -> VersionOracle -> ScriptContext -> Bool

mkVersionOracleValidatorUntyped

Arguments

:: BuiltinData
Genesis UTXO

-> BuiltinData
Datum

-> BuiltinData
Redeemer

-> BuiltinData
ScriptContext

-> BuiltinUnit

compiledVersionOracleValidator :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> BuiltinData -> BuiltinUnit)

serialisableVersionOracleValidator :: SerialisedScript

Version oracle query functions

getVersionedValidatorAddress :: VersionOracleConfig -> VersionOracle -> ScriptContext -> Address

Searches for a specified validator script passed as a reference input. Note that if requested script ID corresponds to a minting policy this function will return a result anyway without reporting any errors.

getVersionedCurrencySymbol :: VersionOracleConfig -> VersionOracle -> ScriptContext -> CurrencySymbol

Searches for a specified minting policy passed as a reference input. Note that if requested script ID corresponds to a validator this function will return a result anyway without reporting any errors.

approvedByGovernance :: VersionOracleConfig -> ScriptContext -> Bool

Check whether a given transaction is approved by Partner Chain governance. The actual check is delegated to a governance minting policy stored in the versioning system. Caller specifies the requested governance version. The transaction must mint at least one token of the governance minting policy to signify transaction approval.

Scripts calling approvedByGovernance must provide the VersionOracle UTXO for governancePolicyId as a reference input.